try to set scan size but not successful
noticed that SPC_get_parameters always returns scan_size_x 512 and
scan_xize_y 2. Therefore we tried to use SPC_set_parameters to set the
scan size to 512x512. however it was not successful as calling
SPC_get_parameters reveals that the scan size remains as 512x2 and
ADC_resolution is still 12. -- Bug and TODO

Also noticed that SPC_get_photons_from_stream always return error -15.
And it seems this fifo streaming method is messed up with the binary fwrite
(which is inside SaveData which is not called in current version) and
therefore need to rewrite the part of the ReadoutLoop function.