start monitoring FLIM parameters as OpenScan starts
original issue with starting monitor loop from the beginning is that if Stop Live is clicked
i.e. stopRequested = true, the monitor thread will exit and there is no way to restart it
without restarting MM

A non-elegant solution:
Add a new parameter monitoringFLIM to FLIM property
by default it is set to True to ensure that the FLIM monitoring loop
keeps running even when Stop Live is clicked.
And the monitoring loop only exits when both a stop is requested (like
Stop Live is clicked or quiting OpenFLIM) and monitoringFLIM is set to
False.