Enabled FLIM implementation and tested on OWSII
with OpenScan version uw-loci/mm-openscan@9e3f691

update FLIM ini file

- able to list FPGA, NI DAQ, and BH SPC150 in Hardware config wizard
and able to combine DAQ (scanner) and SPC150 (detector) to create a
config file

- use DAQ as both scanner and detector appears working fine

- DAQ + BH_SPC150 combination able to run without issue when clicking
'Live'

- tested with real samples and the DAQ+FLIM worked!

Bug:
1. MM crashed when using FPGA as scanner and detector
2. MM crashed when using FPGA as scanner and BH-SPC150 as detector
3. MM crashed when disable FPGA implementation and MM not able to detect
installed DAQ as well
-- we probably wont fix the above bugs as the new OpenScan architecture should solve
the issue

4. BH-SPC150 board is being used whenever OpenScan is running, even when
FLIM implementation is NOT used in the configuration (eg DAQ as scanner
and detector). The problem is caused by EnumerateInstances in current BH_SPC150.c,
as the board is initialized when the function is called by GetInstances
when it searching to list all available devices. -- is it possible to
initialize FLIM board at a later stage after BH_SPC150 is selected in
the config?